misc. cleanup
