+++ rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+profile = "minimal"
+channel = "nightly-2024-09-24"
+components = ["rustfmt", "clippy"]
