+++ rustfmt.toml
@@ -1,4 +1,5 @@
 edition = "2021"
+group_imports = "StdExternalCrate"
 ignore = ["mesh-generated"]
 imports_granularity = "Crate"
 max_width = 120
@@ -7,6 +8,7 @@ overflow_delimited_expr = true
 spaces_around_ranges = true
 tab_spaces = 2
 use_field_init_shorthand = true
+use_small_heuristics = "Max"
 use_try_shorthand = true
 version = "Two"
 wrap_comments = true
