+++ src/api/construction_parse.rs
@@ -1,7 +1,7 @@
-use crate::MinaMesh;
 use anyhow::Result;
-pub use mesh::models::ConstructionParseRequest;
-pub use mesh::models::ConstructionParseResponse;
+pub use mesh::models::{ConstructionParseRequest, ConstructionParseResponse};
+
+use crate::MinaMesh;
 
 /// https://github.com/MinaProtocol/mina/blob/985eda49bdfabc046ef9001d3c406e688bc7ec45/src/app/rosetta/lib/construction.ml#L615
 impl MinaMesh {
