+++ src/api/construction_payloads.rs
@@ -1,7 +1,7 @@
-use crate::MinaMesh;
 use anyhow::Result;
-pub use mesh::models::ConstructionPayloadsRequest;
-pub use mesh::models::ConstructionPayloadsResponse;
+pub use mesh::models::{ConstructionPayloadsRequest, ConstructionPayloadsResponse};
+
+use crate::MinaMesh;
 
 /// https://github.com/MinaProtocol/mina/blob/985eda49bdfabc046ef9001d3c406e688bc7ec45/src/app/rosetta/lib/construction.ml#L473
 impl MinaMesh {
