+++ src/api/mempool_transaction.rs
@@ -1,9 +1,10 @@
-use crate::{
-  graphql::{QueryMempoolTransactions, QueryMempoolTransactionsVariables},
-  MinaMesh, MinaMeshError,
-};
 use cynic::QueryBuilder;
 pub use mesh::models::{MempoolTransactionRequest, MempoolTransactionResponse, Transaction, TransactionIdentifier};
+
+use crate::{
+  MinaMesh, MinaMeshError,
+  graphql::{QueryMempoolTransactions, QueryMempoolTransactionsVariables},
+};
 
 /// https://github.com/MinaProtocol/mina/blob/985eda49bdfabc046ef9001d3c406e688bc7ec45/src/app/rosetta/lib/mempool.ml#L137
 impl MinaMesh {
@@ -11,16 +12,13 @@ impl MinaMesh {
     &self,
     request: MempoolTransactionRequest,
   ) -> Result<MempoolTransactionResponse, MinaMeshError> {
-    let QueryMempoolTransactions {
-      daemon_status: _daemon_status,
-      initial_peers: _initial_peers,
-      pooled_user_commands,
-    } = self
-      .graphql_client
-      .send(QueryMempoolTransactions::build(QueryMempoolTransactionsVariables {
-        hashes: Some(vec![request.transaction_identifier.hash.as_str()]),
-      }))
-      .await?;
+    let QueryMempoolTransactions { daemon_status: _daemon_status, initial_peers: _initial_peers, pooled_user_commands } =
+      self
+        .graphql_client
+        .send(QueryMempoolTransactions::build(QueryMempoolTransactionsVariables {
+          hashes: Some(vec![request.transaction_identifier.hash.as_str()]),
+        }))
+        .await?;
     let operations = pooled_user_commands.into_iter().map(Into::into).collect();
     Ok(MempoolTransactionResponse {
       metadata: None,
