+++ src/graphql/client.rs
@@ -11,10 +11,7 @@ pub struct GraphQLClient {
 
 impl GraphQLClient {
   pub fn new(mina_proxy_url: String) -> Self {
-    Self {
-      mina_proxy_url,
-      client: Client::new(),
-    }
+    Self { mina_proxy_url, client: Client::new() }
   }
 
   pub async fn send<ResponseData, Vars>(
@@ -25,19 +22,9 @@ impl GraphQLClient {
     Vars: serde::Serialize,
     ResponseData: serde::de::DeserializeOwned + 'static,
   {
-    let response = self
-      .client
-      .post(self.mina_proxy_url.to_owned())
-      .run_graphql(operation)
-      .await?;
+    let response = self.client.post(self.mina_proxy_url.to_owned()).run_graphql(operation).await?;
     if let Some(errors) = response.errors {
-      Err(MinaMeshError::GraphqlMinaQuery(
-        errors
-          .into_iter()
-          .map(|err| err.message)
-          .collect::<Vec<_>>()
-          .join("\n\n"),
-      ))
+      Err(MinaMeshError::GraphqlMinaQuery(errors.into_iter().map(|err| err.message).collect::<Vec<_>>().join("\n\n")))
     } else if let Some(data) = response.data {
       Ok(data)
     } else {
