+++ src/graphql/conversions.rs
@@ -1,5 +1,6 @@
-use super::{PublicKey, SyncStatus, UserCommand};
 use mesh::models::{AccountIdentifier, Amount, Currency, Operation, OperationIdentifier, SyncStatus as MeshSyncStatus};
+
+use super::{PublicKey, SyncStatus, UserCommand};
 
 impl From<SyncStatus> for MeshSyncStatus {
   fn from(value: SyncStatus) -> Self {
@@ -11,11 +12,7 @@ impl From<SyncStatus> for MeshSyncStatus {
       SyncStatus::Offline => ("Offline", false),
       SyncStatus::Synced => ("Synced", true),
     };
-    Self {
-      stage: Some(stage.to_string()),
-      synced: Some(synced),
-      ..Default::default()
-    }
+    Self { stage: Some(stage.to_string()), synced: Some(synced), ..Default::default() }
   }
 }
 
@@ -26,10 +23,7 @@ impl From<UserCommand> for Operation {
       r#type: value.kind.0,
       status: Some("pending".to_string()),
       account: Some(Box::new(AccountIdentifier::new(value.source.public_key.0))),
-      amount: Some(Box::new(Amount::new(
-        value.amount.0,
-        Currency::new("mina".to_string(), 9),
-      ))),
+      amount: Some(Box::new(Amount::new(value.amount.0, Currency::new("mina".to_string(), 9)))),
       coin_change: None,
       metadata: None,
       operation_identifier,
