+++ tests/block.rs
@@ -1,22 +1,13 @@
 use anyhow::Result;
-use futures::{stream::FuturesUnordered, StreamExt};
-use mesh::models::partial_block_identifier;
-use mina_mesh::{BlockMetadata, BlockRequest, BlockResponse, NetworkIdentifier, PartialBlockIdentifier, ServeCommand};
-
-async fn setup(partial_block_identifier: PartialBlockIdentifier) -> Result<BlockMetadata> {
-  let mina_mesh = ServeCommand::default().to_mina_mesh().await?;
-  mina_mesh.block_metadata(partial_block_identifier).await
-}
+use futures::{StreamExt, stream::FuturesUnordered};
+use mina_mesh::{BlockMetadata, MinaMeshConfig, PartialBlockIdentifier};
 
 #[tokio::test]
 async fn eq_when_specified() -> Result<()> {
-  let mina_mesh = ServeCommand::default().to_mina_mesh().await?;
+  let mina_mesh = MinaMeshConfig::default().to_mina_mesh().await?;
   let mut metadata_futures = FuturesUnordered::new();
   let partial_block_identifiers = vec![
-    PartialBlockIdentifier {
-      hash: None,
-      index: Some(375991),
-    },
+    PartialBlockIdentifier { hash: None, index: Some(375991) },
     PartialBlockIdentifier {
       hash: Some("3NKAyx2FuWx3jqtkpigndDRoyydQSUozPYix3hw3FWhZc8iUWwTP".to_string()),
       index: None,
@@ -41,11 +32,8 @@ async fn eq_when_specified() -> Result<()> {
 
 #[tokio::test]
 async fn with_neither() -> Result<()> {
-  let result = setup(PartialBlockIdentifier {
-    hash: None,
-    index: None,
-  })
-  .await;
+  let mina_mesh = MinaMeshConfig::default().to_mina_mesh().await?;
+  let result = mina_mesh.block_metadata(PartialBlockIdentifier { hash: None, index: None }).await;
   println!("{:?}", result);
   Ok(())
 }
