--- tests/tests.rs
@@ -1,1 +1,0 @@
-mod account_balance;
